Add support for EnchantScroll (#10)

+++ Maple2.Model/Metadata/Table/ScrollTable.cs
@@ -1,11 +1,12 @@
 ﻿using System.Collections.Generic;
+using Maple2.Model.Enum;
 
 namespace Maple2.Model.Metadata;
 
 public record EnchantScrollTable(IReadOnlyDictionary<int, EnchantScrollMetadata> Entries) : Table;
 
 public record EnchantScrollMetadata(
-    int Type,
+    short Type,
     short MinLevel,
     short MaxLevel,
     int[] Enchants,
